Token to a separate module
