+++ youtube/devdojo-microservices/auth/src/main/java/academy/devdojo/youtube/auth/security/filter/JwtUsernameAndPasswordAuthenticationFilter.java
@@ -2,13 +2,8 @@ package academy.devdojo.youtube.auth.security.filter;
 
 import academy.devdojo.youtube.core.model.ApplicationUser;
 import academy.devdojo.youtube.core.property.JwtConfiguration;
+import academy.devdojo.youtube.security.token.creator.TokenCreator;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.nimbusds.jose.*;
-import com.nimbusds.jose.crypto.DirectEncrypter;
-import com.nimbusds.jose.crypto.RSASSASigner;
-import com.nimbusds.jose.jwk.JWK;
-import com.nimbusds.jose.jwk.RSAKey;
-import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.SignedJWT;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
@@ -17,7 +12,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
@@ -26,24 +20,20 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.security.KeyPair;
-import java.security.KeyPairGenerator;
-import java.security.interfaces.RSAPublicKey;
 import java.util.Collections;
-import java.util.Date;
-import java.util.UUID;
-import java.util.stream.Collectors;
 
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
 @Slf4j
 public class JwtUsernameAndPasswordAuthenticationFilter extends UsernamePasswordAuthenticationFilter {
     private final AuthenticationManager authenticationManager;
     private final JwtConfiguration jwtConfiguration;
+    private final TokenCreator tokenCreator;
 
     @Override
     @SneakyThrows
     public Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response) {
         log.info("Attempting authentication...");
+
         ApplicationUser applicationUser = new ObjectMapper().readValue(request.getInputStream(), ApplicationUser.class);
 
         if (applicationUser == null)
@@ -63,9 +53,9 @@ public class JwtUsernameAndPasswordAuthenticationFilter extends UsernamePassword
     protected void successfulAuthentication(HttpServletRequest request, HttpServletResponse response, FilterChain chain, Authentication auth) throws IOException, ServletException {
         log.info("Authentication was successful for the user '{}', generation JWT token", auth.getName());
 
-        SignedJWT signedJWT = createSignedJWT(auth);
+        SignedJWT signedJWT = tokenCreator.createSignedJWT(auth);
 
-        String encryptedToken = ecryptToken(signedJWT);
+        String encryptedToken = tokenCreator.ecryptToken(signedJWT);
 
         log.info("Token generated successfully, addint it to the response header");
 
@@ -74,80 +64,4 @@ public class JwtUsernameAndPasswordAuthenticationFilter extends UsernamePassword
         response.addHeader(jwtConfiguration.getHeader().getName(), jwtConfiguration.getHeader().getPrefix() + encryptedToken);
     }
 
-    @SneakyThrows
-    private SignedJWT createSignedJWT(Authentication auth){
-        log.info("Starting to create the signed JWT");
-
-       ApplicationUser applicationUser = (ApplicationUser) auth.getPrincipal();
-
-        JWTClaimsSet jwtClaimSet = createJWTClaimSet(auth, applicationUser);
-
-        KeyPair rsaKey = generateKeyPar();
-
-        log.info("Building JWK from the RSA Keys");
-
-        JWK jwk = new RSAKey.Builder((RSAPublicKey) rsaKey.getPublic())
-                .keyID(UUID.randomUUID().toString())
-                .build();
-
-        SignedJWT signedJWT = new SignedJWT(new JWSHeader.Builder(JWSAlgorithm.RS256)
-                .jwk(jwk)
-                .type(JOSEObjectType.JWT)
-                .build(), jwtClaimSet);
-
-        log.info("Signing the token with the private RSA key");
-
-        //chave privada para assinar o token
-        RSASSASigner signer = new RSASSASigner(rsaKey.getPrivate());
-
-        signedJWT.sign(signer);
-
-        log.info("Serialized token '{}'",signedJWT.serialize());
-
-        return signedJWT;
-    }
-
-    private JWTClaimsSet createJWTClaimSet(Authentication auth, ApplicationUser applicationUser){
-        log.info("Creating the JWTClaimSet Object for '{}'", applicationUser);
-        return new JWTClaimsSet.Builder()
-                .subject(applicationUser.getUsername())
-                .claim("authorities", auth.getAuthorities()
-                    .stream()
-                    .map(GrantedAuthority::getAuthority)
-                    .collect(Collectors.toList()))
-                .issuer("http://academy.devdojo")
-                .issueTime(new Date())
-                .expirationTime(new Date(System.currentTimeMillis() + (jwtConfiguration.getExpiration() * 1000)))
-                .build();
-    }
-
-    @SneakyThrows
-    private KeyPair generateKeyPar(){
-        log.info("Generation RS 2018 bits Key");
-
-        KeyPairGenerator generator = KeyPairGenerator.getInstance("RSA");
-
-        generator.initialize(2048);
-
-        return generator.genKeyPair();
-    }
-
-    private String ecryptToken(SignedJWT signedJWT) throws JOSEException {
-        log.info("Starting the ecryptToken method");
-
-        DirectEncrypter directEncrypter = new DirectEncrypter(jwtConfiguration.getPrivateKey().getBytes());
-
-        JWEObject jweObject = new JWEObject(new JWEHeader.Builder(JWEAlgorithm.DIR, EncryptionMethod.A128CBC_HS256)
-                .contentType("JWT")
-                .build(), new Payload(signedJWT));
-
-        log.info("Encrypting token with system's private key");
-
-        jweObject.encrypt(directEncrypter);
-
-        log.info("Token encrypted");
-
-        return jweObject.serialize();
-
-    }
 }
