+++ youtube/devdojo-microservices/auth/src/main/resources/application.yml
@@ -7,7 +7,6 @@ eureka:
   client:
     service-url:
       defaultZone: http://localhost:8081/eureka/
-    fetch-registry: true
     register-with-eureka: true
 
 spring:
