+++ youtube/devdojo-microservices/pom.xml
@@ -22,6 +22,7 @@
     <modules>
         <module>course</module>
         <module>core</module>
+        <module>token</module>
     </modules>
 
     <dependencies>
