+++ youtube/devdojo-microservices/token/src/main/java/academy/devdojo/youtube/security/token/creator/TokenCreator.java
@@ -1,0 +1,112 @@
+package academy.devdojo.youtube.security.token.creator;
+
+import academy.devdojo.youtube.core.model.ApplicationUser;
+import academy.devdojo.youtube.core.property.JwtConfiguration;
+import com.nimbusds.jose.*;
+import com.nimbusds.jose.crypto.DirectEncrypter;
+import com.nimbusds.jose.crypto.RSASSASigner;
+import com.nimbusds.jose.jwk.JWK;
+import com.nimbusds.jose.jwk.RSAKey;
+import com.nimbusds.jwt.JWTClaimsSet;
+import com.nimbusds.jwt.SignedJWT;
+import lombok.RequiredArgsConstructor;
+import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.stereotype.Service;
+
+import java.security.KeyPair;
+import java.security.KeyPairGenerator;
+import java.security.interfaces.RSAPublicKey;
+import java.util.Date;
+import java.util.UUID;
+import java.util.stream.Collectors;
+
+@Service
+@Slf4j
+@RequiredArgsConstructor(onConstructor = @__(@Autowired))
+public class TokenCreator {
+
+    private final JwtConfiguration jwtConfiguration;
+
+    @SneakyThrows
+    public SignedJWT createSignedJWT(Authentication auth){
+        log.info("Starting to create the signed JWT");
+
+        ApplicationUser applicationUser = (ApplicationUser) auth.getPrincipal();
+
+        JWTClaimsSet jwtClaimSet = createJWTClaimSet(auth, applicationUser);
+
+        KeyPair rsaKey = generateKeyPar();
+
+        log.info("Building JWK from the RSA Keys");
+
+        JWK jwk = new RSAKey.Builder((RSAPublicKey) rsaKey.getPublic())
+                .keyID(UUID.randomUUID().toString())
+                .build();
+
+        SignedJWT signedJWT = new SignedJWT(new JWSHeader.Builder(JWSAlgorithm.RS256)
+                .jwk(jwk)
+                .type(JOSEObjectType.JWT)
+                .build(), jwtClaimSet);
+
+        log.info("Signing the token with the private RSA key");
+
+        //chave privada para assinar o token
+        RSASSASigner signer = new RSASSASigner(rsaKey.getPrivate());
+
+        signedJWT.sign(signer);
+
+        log.info("Serialized token '{}'",signedJWT.serialize());
+
+        return signedJWT;
+    }
+
+    private JWTClaimsSet createJWTClaimSet(Authentication auth, ApplicationUser applicationUser){
+        log.info("Creating the JWTClaimSet Object for '{}'", applicationUser);
+
+        return new JWTClaimsSet.Builder()
+                .subject(applicationUser.getUsername())
+                .claim("authorities", auth.getAuthorities()
+                        .stream()
+                        .map(GrantedAuthority::getAuthority)
+                        .collect(Collectors.toList()))
+                .issuer("http://academy.devdojo")
+                .issueTime(new Date())
+                .expirationTime(new Date(System.currentTimeMillis() + (jwtConfiguration.getExpiration() * 1000)))
+                .build();
+    }
+
+    @SneakyThrows
+    private KeyPair generateKeyPar(){
+        log.info("Generation RS 2018 bits Key");
+
+        KeyPairGenerator generator = KeyPairGenerator.getInstance("RSA");
+
+        generator.initialize(2048);
+
+        return generator.genKeyPair();
+    }
+
+    public String ecryptToken(SignedJWT signedJWT) throws JOSEException {
+        log.info("Starting the ecryptToken method");
+
+        DirectEncrypter directEncrypter = new DirectEncrypter(jwtConfiguration.getPrivateKey().getBytes());
+
+        JWEObject jweObject = new JWEObject(new JWEHeader.Builder(JWEAlgorithm.DIR, EncryptionMethod.A128CBC_HS256)
+                .contentType("JWT")
+                .build(), new Payload(signedJWT));
+
+        log.info("Encrypting token with system's private key");
+
+        jweObject.encrypt(directEncrypter);
+
+        log.info("Token encrypted");
+
+        return jweObject.serialize();
+
+    }
+
+}
